Added neighbor joining file w/ changes
So I made 3 functions, one for calculating naiive prob of residues observed for whole data set, one for getting the columns, and one for calculating probability of residue in each column. These are of course subject to change, especially the formatting because I didn't want to use pandas much for some reason. Feel free to experiment further.