New equation attempt
Hey guys so I have been trying to implement equation 1 from the reading. I have a simple rundown of how it works but am having issues with the matrix creation or something as we get the same numbers over and over. Look at my comments and let me know if anything needs explanation.